Implementing a shiftwidth of 4
Before, the shiftwidth was almost random. Regularizing this to any number (it
doesn't have to be 4, that's just my arbitrary preference) will go a long way to
code readability and attracting people to the project. By not scaring them away
because they don't know what `gg=G` means in VIM.